Fix bugs
First off, the program would only load the first plugin in the plugins
list. Turns out that this was caused by enquoting the "plugins"
variable. Fix that.

Second, a variable wasn't properly assigned, due to poorly used
quotation marks, which lead to all sorts of nasty bugs. Fix that as
well.